Updates to the server
Testing performed with one client
receiving data from the client is working properly
sending data to the client is working properly

Strange behavior: The client is printing the entire transaction history.
I flush the output to the server before sending the data back, so I
believe this is on the client side.